<refactor> : 리팩토링에 따른 응답 DTO들 수정
