+++ src/main/java/com/example/demo/domain/dto/response/Member/RegisterResponse.java
@@ -1,9 +1,10 @@
 package com.example.demo.domain.dto.response.Member;
 
-import com.example.demo.domain.dto.response.ApiResponse;
+import com.example.demo.support.ApiResponse;
 
-public class RegisterResponse extends ApiResponse {
-    public RegisterResponse(String message, int status, boolean success) {
-        super(message, status, success);
+public class RegisterResponse  {
+
+    public RegisterResponse() {
+
     }
 }
